// richardhan-server/src/common/constants/enums.ts
export enum VisibilityEnum {
  PUBLIC = 'PUBLIC',
  FRIENDS = 'FRIENDS',
  PRIVATE = 'PRIVATE',
}

export enum ContentTypeEnum {
  POST = 'POST',
  REEL = 'REEL',
}

export enum GenderEnum {
  MALE = 'MALE',
  FEMALE = 'FEMALE',
}

export enum VaccinationStatusEnum {
  UPDATED = 'UPDATED',
  NOT_UPDATED = 'NOT_UPDATED',
  DUE = 'DUE',
}

export enum PetSitterBookingTypeEnum {
  PACKAGE = 'PACKAGE',
  SERVICE = 'SERVICE',
}

export enum FriendshipRelationStatusEnum {
  FRIENDS = 'FRIENDS',
  UNFRIENDED = 'UNFRIENDED',
  BLOCKED = 'BLOCKED',
  REQUEST_SENT = 'REQUEST_SENT',
  REQUEST_RECEIVED = 'REQUEST_RECEIVED',
  NONE = 'NONE',
  SELF = 'SELF',
}